move logic to init API data stores with data from apiData to storeMocks
update ProgramSelector tests to use this logic instead
